pyproject.toml
[project]
name = "plugins"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.5.0",
    "hydra-core>=1.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."ds_plugins"]
plugin_1 = "plugins.plugin1:Plugin1"
plugin_2 = "plugins.plugin2:Plugin2"
plugin_3 = "plugins.plugin3:Plugin3"